apps/api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
httpx==0.28.1
pydantic==2.11.4
python-dotenv==1.1.0
psycopg[binary,pool]==3.2.9
python-jose[cryptography]==3.4.0
cryptography==44.0.3
